requirements.txt
streamlit==1.33.0
pandas==2.2.2
python-dotenv==1.0.1
google-generativeai==0.8.4
serpapi==1.7.0
plotly==5.22.0
